+++ e2e-tests/development-runtime/src/pages/client-dynamic-route/[id].js
@@ -1,9 +1,11 @@
 import React from "react"
 import Layout from "../../components/layout"
 
-export default props => (
+const ClientDynamicID = props => (
   <Layout>
     <h1 data-testid="title">Client Dynamic Route</h1>
     <h2 data-testid="params">{props.params.id}</h2>
   </Layout>
 )
+
+export default ClientDynamicID